MVMRcML method with Data Perturbation

This is the internal MVMRcML-BIC function of mr_mvcML.

Usage

MVmr_cML(
  b_exp,
  b_out,
  se_bx,
  Sig_inv_l,
  n,
  K_vec = as.numeric(c()),
  random_start = 1L,
  min_theta_range = -0.5,
  max_theta_range = 0.5,
  maxit = 100L,
  thres = 1e-04
)

Arguments

b_exp

A m*L matrix of SNP effects on the exposure variable.

b_out

A m*1 matrix of SNP effects on the outcome variable.

se_bx

A m*L matrix of standard errors of b_exp.

Sig_inv_l

A list of the inverse of m covariance matrices.

n

The smallest sample size of the L+1 GWAS dataset.

K_vec

Sets of candidate K's, the constraint parameter representing number of invalid IVs.

random_start

Number of random start points, default is 1.

min_theta_range

The lower bound of the uniform distribution for each initial value for theta generated from.

max_theta_range

The upper bound of the uniform distribution for each initial value for theta generated from.

maxit

Maximum number of iterations for each optimization, default is 100.

thres

Threshold for convergence criterion.

Value

A list

BIC_theta

Estimated causal effect from MVMR-cML-BIC

BIC_invalid

Invalid IVs selected by MVMR-cML-BIC

l_vec

A vector of negative log-likelihood corresponding to each K.

K_vec

A vector of candidate K's

theta_vec

A matrix of causal parameter estimates, each column corresponds to a candidate K.

Conv_vec

A vector of successful convergence indicators corresponding to each K.

Converge

Indicator of successful convergence, 0 means success, 1 means failure.

BIC_vec

Data perturbation with successful convergence

Khat

The length of BIC_invalid.