Conditional likelihood ratio (CLR) method

The mr_clr function calculates confidence intervals based on inverting the conditional likelihood ratio and other identification-robust tests.

Usage

mr_clr(object, nx, ny, alpha = 0.05, CIMin = -10, CIMax = 10, CIStep = 0.01)

# S4 method for MRInput
mr_clr(object, nx, ny, alpha = 0.05, CIMin = -10, CIMax = 10, CIStep = 0.01)

Arguments

object

An MRInput object.

nx

The sample size used to compute genetic associations with the exposure.

ny

The sample size used to compute genetic associations with the outcome.

alpha

The significance level used to calculate the confidence interval. The default value is 0.05.

CIMin

The smallest value to use in the search to find the confidence interval (default is -10).

CIMax

The largest value to use in the search to find the confidence interval (default is +10).

CIStep

The step size to use in the search to find the confidence interval (default is 0.01). Using a lower value (such as 0.001) will give more precise confidence intervals, but increase run time.

Value

The output from the function is an CLR object containing:

Exposure

A character string with the name given to the exposure.

Outcome

A character string with the names given to the outcome.

Correlation

The matrix of genetic correlations.

ARlower

The lower bounds of the causal estimate based on inverting Anderson and Rubin's test.

ARupper

The upper bounds of the causal estimate based on inverting Anderson and Rubin's test.

Klower

The lower bounds of the causal estimate based on inverting Kleibergen's test.

Kupper

The upper bounds of the causal estimate based on inverting Kleibergen's test.

CLRlower

The lower bounds of the causal estimate based on inverting Moreira's conditional likelihood ratio test.

CLRupper

The upper bounds of the causal estimate based on inverting Moreira's conditional likelihood ratio test.

CIMin

The smallest value used in the search to find the confidence interval.

CIMax

The largest value used in the search to find the confidence interval.

CIStep

The step size used in the search to find the confidence interval.

Alpha

The significance level used when calculating the confidence intervals.

Details

In weak instrument settings, usual inference based on point estimates and standard errors may not be accurate. This method calculates confidence intervals based on inverting identification-robust tests proposed in Wang and Kang (2021, Biometrics) that provide valid inferences regardless of instrument strength.

This includes conditional likelihood ratio (CLR), Kleibergen (K), and Anderson and Rubin (AR) tests.

Evidence from the econometrics literature suggests that CLR inference is the best option in terms of power under a wide range of settings.

Please note that these methods do not provide point estimates, only confidence intervals. While most examples provide a confidence interval that is a single range of values, in some cases the confidence interval may comprise multiple ranges of values. In other cases, a valid confidence interval may not exist.

References

Description of the CLR method: "Weak-instrument robust tests in two-sample summary-data Mendelian randomization", S. Wang and H. Kang, Biometrics, 2021.

Examples

mr_clr(mr_input(bx = calcium, bxse = calciumse,
   by = fastgluc, byse = fastglucse, correl = calc.rho), nx=6351, ny=133010)
#> 
#> --------------------------------------------------------------------------
#> 
#>  95 % confidence intervals using identification-robust methods 
#> 
#> Anderson--Rubin:              [  0.16 ,  5.30 ]  
#> Kleibergen:                   [  0.98 ,  3.88 ]  
#> Conditional likelihood ratio: [  0.96 ,  3.92 ]  
#> --------------------------------------------------------------------------