MR-Lasso method

The mr_lasso function performs the MR-Lasso method, which applies lasso-type penalization to the direct effects of genetic variants on the outcome. The causal estimate is described as a post-lasso estimate, and is obtained by performing the IVW method using only those genetic variants that are identified as valid by the lasso procedure.

Usage

mr_lasso(object, distribution = "normal", alpha = 0.05, lambda = numeric(0))

# S4 method for MRInput
mr_lasso(object, distribution = "normal", alpha = 0.05, lambda = numeric(0))

Arguments

object

An MRInput object.

distribution

The type of distribution used to calculate the confidence intervals. Options are "normal" (default) or "t-dist".

alpha

The significance level used to calculate the confidence intervals. The default value is 0.05.

lambda

The value of the tuning parameter used by the lasso procedure which controls the level of sparsity. If not specified, the tuning parameter will be calculated by the heterogeneity stopping rule.

Value

The output from the function is an MRLasso object containing:

Exposure

A character vector with the names given to the exposure.

Outcome

A character string with the names given to the outcome.

Estimate

The causal estimate from the MR-Lasso method. This is the post-lasso estimate.

StdError

The standard error of the causal estimate from the MR-Lasso method.

CILower

The lower bound of the confidence interval for the causal estimate based on the estimated standard error and the significance level provided.

CIUpper

The upper bound of the confidence interval for the causal estimate based on the estimated standard error and the significance level provided.

Alpha

The significance level used when calculating the confidence intervals.

Pvalue

The p-value associated with the causal estimate using a normal or t-distribution (as specified in distribution).

SNPs

The number of genetic variants (SNPs) included in the analysis.

RegEstimate

The estimate from the regularized regression model used in the MR-Lasso method.

RegIntercept

The intercept estimates from the regularized regression model used in the MR-Lasso method.

Valid

The number of genetic variants that have been identified as valid instruments.

ValidSNPs

The names of genetic variants that have been identified as valid instruments.

Lambda

The value of the tuning parameter used to compute RegEstimate

Details

MR-Lasso extends the IVW model to include an intercept term for each genetic variant. These intercept terms represent associations between the genetic variants and the outcome which bypass the risk factor. The causal effect estimates are estimated by weighted linear regression where the intercept terms are subject to lasso-type penalization. The lasso penalization will tend to shrink the intercept terms corresponding to the valid instruments to zero.

The lasso penalty relies on a tuning parameter which controls the level of sparsity. The default is to use a heterogeneity stopping rule, but a fixed value may be specified.

As part of the analysis, the genetic variants are orientated so that all of the associations with the risk factor are positive (and signs of associations with the outcome are changed to keep the orientation consistent if required). Re-orientation of the genetic variants is performed automatically as part of the function.

The MR-Lasso method is performed in two steps. First, a regularized regression model is fitted, and some genetic variants are identified as valid instruments. Second, the causal effect is estimated using standard IVW with only the valid genetic variants. The post-lasso method will be performed as long as at least two genetic variants are identified as valid instruments. The default heterogeneity stopping rule will always return at least two genetic variants as valid instruments. The main estimate given by the method is the post-lasso estimate. However, parameter estimates from the regularized regression model used to identify invalid variants are also provided for completeness.

If a substantial proportion of genetic variants are removed from the analysis, the MR-Lasso method may give a false impression of confidence in the causal estimate due to homogeneity of the variant-specific causal estimates amongst the remaining variants. However, it is not reasonable to claim that there is strong evidence for a causal effect after a large number of variants with heterogeneous estimates have been removed from the analysis.

References

Jessica MB Rees, Angela M Wood, Frank Dudbridge, Stephen Burgess. Robust methods in Mendelian randomization via penalization of heterogeneous causal estimates. PLoS ONE 2019; 14(9):e0222362

Examples

mr_lasso(mr_input(bx = ldlc, bxse = ldlcse, by = chdlodds, byse = chdloddsse))
#> 
#> MR-Lasso method 
#> 
#> Number of variants : 28 
#> Number of valid instruments : 26 
#> Tuning parameter : 0.7757552 
#> ------------------------------------------------------------------
#>  Exposure Estimate Std Error 95% CI       p-value
#>  exposure    2.671     0.431 1.827, 3.515   0.000
#> ------------------------------------------------------------------