Multivariable MR-Egger method

The mr_mvegger function performs multivariable Mendelian randomization via the MR-Egger method. This is implemented by multivariable weighted linear regression.

Usage

mr_mvegger(
  object,
  orientate = 1,
  correl = FALSE,
  distribution = "normal",
  alpha = 0.05
)

# S4 method for MRMVInput
mr_mvegger(
  object,
  orientate = 1,
  correl = FALSE,
  distribution = "normal",
  alpha = 0.05
)

Arguments

object

An MRMVInput object.

orientate

The risk factor that genetic associations are orientated to. The univariable and multivariable versions of MR-Egger are both sensitive to the choice of parameterization of the genetic associations - which allele the associations are orientated with respect to (in other words, which allele is the effect allele). For univariable MR-Egger, this is resolved by setting the genetic associations with the exposure all to be positive. In multivariable MR-Egger, we have to choose which of the exposures to orientate the genetic associations to. The default option is 1, meaning that genetic associations with the first exposure are set to be positive.

correl

If the genetic variants are correlated, then this correlation can be accounted for. The matrix of correlations between must be provided in the MRInput object: the elements of this matrix are the correlations between the individual variants (diagonal elements are 1). If a correlation matrix is specified in the MRInput object, then correl is set to TRUE.

distribution

The type of distribution used to calculate the confidence intervals. Options are "normal" (default) or "t-dist".

alpha

The significance level used to calculate the confidence interval. The default value is 0.05.

Value

The output from the function is an MVEgger object containing:

Model

A character string giving the type of model used ("random").

Orientate

The number corresponding to the risk factor that the genetic associations are orientated to.

Exposure

A character vector with the names given to the exposure.

Outcome

A character string with the names given to the outcome.

Correlation

The matrix of genetic correlations.

Estimate

A vector of the causal estimates (slope coefficient).

StdError.Est

Standard errors of the causal estimates.

Pvalue.Est

The p-values associated with the estimates using a normal or t-distribution (as specified in distribution).

CILower.Est

The lower bound of the causal estimates based on the estimated standard error and the significance level provided.

CIUpper.Est

The upper bound of the causal estimates based on the estimated standard error and the significance level provided.

Intercept

The value of the intercept estimate.

StdError.Int

Standard error of the intercept estimate.

Pvalue.Int

The p-value associated with the intercept.

CILower.Int

The lower bound of the intercept based on the estimated standard error and the significance level provided.

CIUpper.Int

The upper bound of the intercept based on the estimated standard error and the significance level provided.

Alpha

The significance level used when calculating the confidence intervals.

Pvalue

The p-values associated with the estimates (calculated as Estimate/StdError as per Wald test) using a normal or t-distribution (as specified in distribution).

SNPs

The number of genetic variants (SNPs) included in the analysis.

RSE

The estimated residual standard error from the regression model.

Heter.Stat

Heterogeneity statistic (Cochran's Q statistic) and associated p-value: the null hypothesis is that all genetic variants estimate the same causal parameter; rejection of the null is an indication that one or more variants may be pleiotropic.

Details

Multivariable MR-Egger is an extension of the MR-Egger method to deal with genetic variants that are associated with multiple risk factors.

We implement the method using multivariable weighted linear regression. If the variants are correlated, the method is implemented using generalized weighted linear regression; this is hard coded using matrix algebra.

The causal estimate is obtained by regression of the associations with the outcome on the associations with the risk factors, with the intercept estimated and weights being the inverse-variances of the associations with the outcome.

References

Jessica Rees, Angela Wood, Stephen Burgess. Extending the MR-Egger method for multivariable Mendelian randomization to correct for both measured and unmeasured pleiotropy. Statistics in Medicine 2017; 36(29): 4705-4718. doi: 10.1002/sim.7492.

Examples

mr_mvegger(mr_mvinput(bx = cbind(ldlc, hdlc, trig), bxse = cbind(ldlcse, hdlcse, trigse),
   by = chdlodds, byse = chdloddsse), orientate = 1)
#> 
#> Multivariable MR-Egger method
#> (variants uncorrelated, random-effect model)
#> 
#> Orientated to exposure : 1 
#> Number of Variants : 28 
#> ------------------------------------------------------------------
#>     Exposure Estimate Std Error  95% CI        p-value
#>   exposure_1    2.829     0.515  1.819,  3.839   0.000
#>   exposure_2   -0.663     0.496 -1.636,  0.310   0.182
#>   exposure_3    0.827     0.209  0.417,  1.237   0.000
#>  (intercept)   -0.028     0.010 -0.049, -0.008   0.007
#> ------------------------------------------------------------------
#> Residual standard error =  1.280 
#> Heterogeneity test statistic = 39.3421 on 24 degrees of freedom, (p-value = 0.0251)