Multivariable generalized method of moments (GMM) method

The mr_mvgmm function performs multivariable Mendelian randomization via the generalized method of moments method.

Usage

mr_mvgmm(object, nx, ny, cor.x = NULL, robust = TRUE, alpha = 0.05, ...)

# S4 method for MRMVInput
mr_mvgmm(object, nx, ny, cor.x = NULL, robust = TRUE, alpha = 0.05, ...)

Arguments

object

An MRMVInput object.

nx

Vector of sample sizes used to compute genetic associations with the exposure (one for each exposure).

ny

The sample size used to compute genetic associations with the outcome.

cor.x

Correlation matrix for exposures. Default is to assume the exposures are uncorrelated.

robust

Indicates whether overdispersion heterogeneity is accounted for in the model. Default is TRUE.

alpha

The significance level used to calculate the confidence interval. The default value is 0.05.

...

Additional arguments to be passed to the optimization routines to calculate GMM estimates and overdispersion parameter.

Value

The output from the function is an MVGMM object containing:

Robust

TRUE if overdispersion heterogeneity was included in the model, FALSE otherwise.

Exposure

A character vector with the names given to the exposure.

Outcome

A character string with the names given to the outcome.

Correlation

The matrix of genetic correlations if supplied. If not supplied, then an identity matrix will be returned.

ExpCorrelation

TRUE if an exposure correlation matrix was specified, FALSE otherwise.

CondFstat

A vector of conditional F-statistics (one for each exposure).

Estimate

A vector of causal estimates.

StdError

A vector of standard errors of the causal estimates.

CILower

The lower bounds of the causal estimates based on the estimated standard errors and the significance level provided.

CIUpper

The upper bounds of the causal estimates based on the estimated standard errors and the significance level provided.

Overdispersion

The estimate of the overdispersion parameter. If this is negative, then a value of zero is used in the method.

Pvalue

The p-values associated with the estimates (calculated as Estimate/StdError as per Wald test) using a normal distribution.

Alpha

The significance level used when calculating the confidence intervals.

Heter.Stat

Heterogeneity statistic (Cochran's Q statistic) and associated p-value (for non-robust model only): the null hypothesis is that all genetic variants estimate the same causal parameter; rejection of the null is an indication that one or more genetic variants may be pleiotropic. If the robust option is set to TRUE, then this represents excess heterogeneity beyond the overdispersion model.

Details

Robust inference in two-sample multivariable Mendelian randomization using the generalized method of moments. The method accounts for overdispersion heterogeneity in genetic variant-outcome associations.

References

Description of the generalized method of moments: Hansen, L. P. (1982). Large sample properties of generalized method of moments estimators. Econometrica, pp.1029-1054.

Examples

mr_mvgmm(mr_mvinput(bx = cbind(ldlc, hdlc, trig), bxse = cbind(ldlcse, hdlcse, trigse),
   by = chdlodds, byse = chdloddsse), nx=rep(17723,3), ny=17723)
#> 
#> Multivariable generalized method of moments (GMM) method
#> 
#> Exposure correlation matrix not specified. Exposures are assumed to be uncorrelated.
#> 
#> Robust model with overdispersion heterogeneity.
#> 
#> ------------------------------------------------------------------
#>    Exposure Estimate Std Error  95% CI       p-value Cond F-stat
#>  exposure_1    1.830     0.451  0.947, 2.713   0.000        20.3
#>  exposure_2   -0.722     0.580 -1.858, 0.415   0.213        12.9
#>  exposure_3    0.686     0.239  0.217, 1.155   0.004        13.3
#> ------------------------------------------------------------------
#> 
#> Overdispersion heterogeneity parameter estimate = 17.369 
#> 
#> Heterogeneity test statistic = 24.9273