Multivariable inverse-variance weighted method

The mr_mvivw function performs multivariable Mendelian randomization via the inverse-variance method. This is implemented by multivariable weighted linear regression.

Usage

mr_mvivw(
  object,
  model = "default",
  robust = FALSE,
  correl = FALSE,
  correl.x = NULL,
  nx = NA,
  distribution = "normal",
  alpha = 0.05,
  ...
)

# S4 method for MRMVInput
mr_mvivw(
  object,
  model = "default",
  robust = FALSE,
  correl = FALSE,
  correl.x = NULL,
  nx = NA,
  distribution = "normal",
  alpha = 0.05,
  ...
)

Arguments

object

An MRMVInput object.

model

What type of model should be used: "default", "random" or "fixed". The random-effects model ("random") is a multiplicative random-effects model, allowing overdispersion in the weighted linear regression (the residual standard error is not fixed to be 1, but is not allowed to take values below 1). The fixed-effect model ("fixed") sets the residual standard error to be 1. The "default" setting is to use a fixed-effect model with 3 genetic variants or fewer, and otherwise to use a random-effects model.

robust

Indicates whether robust regression using the lmrob() function from the package robustbase should be used in the method rather than standard linear regression (lm).

correl

If the genetic variants are correlated, then this correlation can be accounted for. The matrix of correlations between must be provided in the MRMVInput object: the elements of this matrix are the correlations between the individual variants (diagonal elements are 1). If a correlation matrix is specified in the MRMVInput object, then correl is set to TRUE.

correl.x

Correlation matrix for exposures (Optional). Default is to assume the exposures are uncorrelated. This is only used in the computation of conditional F-statistics.

nx

Either a single value, or a vector of sample sizes for the genetic associations with the exposures (one for each exposure, and assumed equal for all variants). If a single value is provided, it is assumed this is the sample size for all exposures. This is optional, and it is only used in the calculation of conditional F statistics. If not supplied, then conditional F statistics are not reported.

distribution

The type of distribution used to calculate the confidence intervals. Options are "normal" (default) or "t-dist".

alpha

The significance level used to calculate the confidence interval. The default value is 0.05.

...

Additional arguments to be passed to the regression method.

Value

The output from the function is an MVIVW object containing:

Model

A character string giving the type of model used ("fixed", "random", or "default").

Exposure

A character vector with the names given to the exposure.

Outcome

A character string with the names given to the outcome.

Robust

TRUE if robust regression has been used to calculate the estimate, FALSE otherwise.

Correlation

The matrix of genetic correlations.

Estimate

A vector of causal estimates.

StdError

A vector of standard errors of the causal estimates.

CILower

The lower bounds of the causal estimates based on the estimated standard errors and the significance level provided.

CIUpper

The upper bounds of the causal estimates based on the estimated standard errors and the significance level provided.

Alpha

The significance level used when calculating the confidence intervals.

Pvalue

The p-values associated with the estimates (calculated as Estimate/StdError as per Wald test) using a normal or t-distribution (as specified in distribution).

SNPs

The number of genetic variants (SNPs) included in the analysis.

RSE

The estimated residual standard error from the regression model.

Heter.Stat

Heterogeneity statistic (Cochran's Q statistic) and associated p-value: the null hypothesis is that all genetic variants estimate the same causal parameter; rejection of the null is an indication that one or more variants may be pleiotropic.

CondFstat

Conditional F statistics: An approximation of the first-stage conditional F statistics for all variants based on the summarized data. This represents the instrument strength for each exposure conditional on other exposures in the model. This is only reported if the sample sizes for the genetic associations with the exposures are provided.

Details

Multivariable Mendelian randomization is an extension of Mendelian randomization to deal with genetic variants that are associated with multiple risk factors. Two scenarios are envisioned for its use: 1) risk factors that are biologically related, such as lipid fractions; and 2) risk factors where there is potentially a network of causal effects (mediation) from one risk factor to another. In both cases, under the extended assumptions of multivariable Mendelian randomization, coefficients represent the direct causal effects of each risk factor in turn with the other risk factors being fixed.

We implement the method using multivariable weighted linear regression. If the variants are correlated, the method is implemented using generalized weighted linear regression; this is hard coded using matrix algebra.

The causal estimate is obtained by regression of the associations with the outcome on the associations with the risk factors, with the intercept set to zero and weights being the inverse-variances of the associations with the outcome.

References

Description of approach: Stephen Burgess, Simon G Thompson. Multivariable Mendelian Randomization: the use of pleiotropic genetic variants to estimate causal effects. American Journal of Epidemiology 2015; 181(4):251-260. doi: 10.1093/aje/kwu283.

Description of inverse-variance weighted method: Stephen Burgess, Frank Dudbridge, Simon G Thompson. Re: "Multivariable Mendelian randomization: the use of pleiotropic genetic variants to estimate causal effects." American Journal of Epidemiology 2015; 181(4):290-291. doi: 10.1093/aje/kwv017.

Use for mediation analysis: Stephen Burgess, Deborah J Thompson, Jessica MB Rees, Felix R Day, John R Perry, Ken K Ong. Dissecting causal pathways using Mendelian randomization with summarized genetic data: Application to age at menarche and risk of breast cancer. Genetics 2017; 207(2):481-487. doi: 10.1534/genetics.117.300191.

Calculation of conditional F statistics: Ashish Patel, Dipender Gill, Dmitry Shungin, Christos Mantzoros, Lotte Bjerre Knudsen, Jack Bowden, Stephen Burgess. Robust use of phenotypic heterogeneity at drug target genes for mechanistic insights: application of multivariable Mendelian randomization to GLP1R gene region. Pre-print.

Examples

mr_mvivw(mr_mvinput(bx = cbind(ldlc, hdlc, trig), bxse = cbind(ldlcse, hdlcse, trigse),
   by = chdlodds, byse = chdloddsse))
#> 
#> Multivariable inverse-variance weighted method
#> (variants uncorrelated, random-effect model)
#> 
#> Number of Variants : 28 
#> 
#> ------------------------------------------------------------------
#>    Exposure Estimate Std Error  95% CI       p-value
#>  exposure_1    1.925     0.439  1.064, 2.786   0.000
#>  exposure_2   -0.590     0.555 -1.677, 0.498   0.288
#>  exposure_3    0.723     0.230  0.272, 1.174   0.002
#> ------------------------------------------------------------------
#> Residual standard error =  1.433 
#> Heterogeneity test statistic = 51.3599 on 25 degrees of freedom, (p-value = 0.0014)