Calculates p-values for penalization of weights with second-order weights

Internal function for calculating penalized weights in conjunction with r.weights with weights equal to "delta".

These weights are used in either the mr_ivw or mr_egger functions when penalized = TRUE, or in the mr_median function when method = "penalized".

Usage

penalised.weights.delta(.Bx, .Bxse, .By, .Byse, .psi)

Arguments

.Bx

Beta-coefficient for genetic association with the risk factor.

.Bxse

Standard error of genetic association with the risk factor.

.By

Beta-coefficient for genetic association with the outcome.

.Byse

Standard error of genetic association with the outcome.

.psi

The correlation between the genetic associations with the exposure and the association with the outcome for each variant resulting from sample overlap.

Value

P-values corresponding to heterogeneity test for each genetic variant in turn (based on a chi-squared-1 distribution).

Details

None.

Examples

penalised.weights.delta(ldlc, ldlcse, chdlodds, chdloddsse, 0)
#>  [1] 0.46846756 0.01077052 0.23978581 0.40041370 0.04433067 0.30223393
#>  [7] 0.81043446 0.94637358 0.09900936 0.25362247 0.73904568 0.30610466
#> [13] 0.43695043 0.00438833 0.62700442 0.60761702 0.64599054 0.35582637
#> [19] 0.13695840 0.04083391 0.97578152 0.38790748 0.10655035 0.92101086
#> [25] 0.07918300 0.18173663 0.53561298 0.20042314